Refactored FOV to allow for abstraction of the calculation method.
